+++ src/main.c
@@ -58,6 +58,7 @@ struct Config config = {
     "Kernel",   // kernel_prefix
     "Desktop",  // desktop_prefix
     "Theme",    // gtk_theme_prefix
+    "Icons",    // icon_theme_prefix
     "Shell",    // shell_prefix
     "Login",    // login_shell_prefix
     "Terminal", // term_prefix
@@ -406,6 +407,7 @@ int main(int argc, char **argv) {
         {"kernel", config.kernel_prefix, kernel},
         {"desktop", config.desktop_prefix, desktop},
         {"gtk_theme", config.gtk_theme_prefix, gtk_theme},
+        {"icon_theme", config.icon_theme_prefix, icon_theme},
         {"shell", config.shell_prefix, shell},
         {"login_shell", config.login_shell_prefix, login_shell},
         {"term", config.term_prefix, term},