+++ src/utils.h
@@ -60,6 +60,7 @@ struct Config {
     char kernel_prefix[64];
     char desktop_prefix[64];
     char gtk_theme_prefix[64];
+    char icon_theme_prefix[64];
     char shell_prefix[64];
     char login_shell_prefix[64];
     char term_prefix[64];